refactor(playback-target): orchestrate MPV events and progress in one module
Move MPV property, end-file, client-message, Intro Skipper,
progress-reporting, and disconnect handling out of SessionManager into
one internal Playback Target event module. The loop owns when to
refresh but delegates state projection and emission to the shared Now
Playing owner, so SessionManager carries no second Now Playing builder
and no new Tauri event adapter. Pause, volume, and mute still report
immediately while position stays throttled to five seconds;
automatic, manual, and off Intro Skipper behavior with once-per-range
rules is unchanged; natural end still advances to the next episode
while Credit Skip only seeks; receiver closure clears the playback
context and emits the shared disconnected Now Playing state. Event
names and payloads are unchanged, and the existing event, Intro
Skipper, progress, and Emby cleanup tests stay green against the new
module paths.

Closes #211
